\name{hmcpkg-package}
\alias{hmcpkg-package}
\alias{hmcpkg}
\docType{package}
\title{
\packageTitle{hmcpkg}
}
\description{
\packageDescription{hmcpkg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hmcpkg}
\packageIndices{hmcpkg}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{hmcpkg}

Maintainer: \packageMaintainer{hmcpkg}
}
\references{
Hogg, R.V., McKean, J.W. and Craig, A.T. (2018), \emph{Introduction to Mathematical Statistics, 8th Ed.}, Pearson: Boston.

~~ Literature or other references for background information ~~
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
